requirements.txt
transformers
torch
sentencepiece
sentence-transformers
pymupdf
python-docx
pandas
streamlit
